# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a OpenAPI 2 specification for Mashrooy"
  version: "1.0.0"
  title: "OpenAPI Mashrooy"
host: "localhost:3006"
basePath: "/Mashrooy"
...
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /product:
    get:
      tags:
      - "Product"
      summary: "list Products"
      operationId: "listProduct"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "rps"
        in: "query"
        description: "Size of the page results in pagination"
        required: false
        type: "integer"
      - name: "rpi"
        in: "query"
        description: "Current page index in pagination"
        required: false
        type: "integer"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Product"
        400:
          description: "Invalid username supplied"
          schema:
            type: "string"
        403:
          description: "Forbidden"
          schema:
            type: "string"
        404:
          description: "Products not found"
          schema:
            type: "string"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "Product"
    post:
      tags:
      - "Product"
      summary: "Add new product"
      operationId: "creatProduct"
      parameters:
      - in: "body"
        name: "product"
        description: "product array that needs to be added"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/Product"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Product"
        400:
          description: "Validation Error"
          schema:
            type: "string"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ApiResponse"
        409:
          description: "Conflict record"
          schema:
            $ref: "#/definitions/ApiResponse"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "Product"
  /product/uploadProductAttachments:
    post:
      tags:
      - "Product"
      summary: "Upload Products."
      operationId: "uploadProductAttachments"
      consumes:
      - "multipart/form-data"
      parameters:
      - name: "upfile"
        in: "formData"
        description: "The image to upload."
        required: false
        type: "file"
      responses:
        200:
          description: "successful operation"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "Product"
  /product/{id}:
    get:
      tags:
      - "Product"
      summary: "Get Product by Id"
      operationId: "getProductById"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of the product that needs to be fetched."
        required: true
        type: "string"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Product"
        400:
          description: "Invalid username supplied"
          schema:
            type: "string"
        403:
          description: "Forbidden"
          schema:
            type: "string"
        404:
          description: "Products not found"
          schema:
            type: "string"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "Product"
    put:
      tags:
      - "Product"
      summary: "Updated Product"
      description: "This can only be done by the logged in employee."
      operationId: "updateProductById"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id"
        in: "path"
        description: "The id of the product that needs to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated Product object"
        required: true
        schema:
          $ref: "#/definitions/Product"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Product"
        400:
          description: "Invalid username supplied"
          schema:
            type: "string"
        403:
          description: "Forbidden"
          schema:
            type: "string"
        404:
          description: "Products not found"
          schema:
            type: "string"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "Product"
    delete:
      tags:
      - "Product"
      summary: "Delete Product"
      description: "This can only be done by the logged in employee."
      operationId: "deleteProductById"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The Product id that needs to be deleted"
        required: true
        type: "string"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Product"
        400:
          description: "Invalid username supplied"
          schema:
            type: "string"
        403:
          description: "Forbidden"
          schema:
            type: "string"
        404:
          description: "Products not found"
          schema:
            type: "string"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "Product"
  /product/listdbyFilters:
    post:
      tags:
      - "Product"
      summary: "List Products By specific Filters"
      operationId: "listProductsByFilters"
      parameters:
      - name: "rps"
        in: "query"
        description: "Size of the page results in pagination"
        required: false
        type: "integer"
      - name: "rpi"
        in: "query"
        description: "Current page index in pagination"
        required: false
        type: "integer"
      - in: "body"
        name: "matchingParams"
        description: "The params to fetch users."
        required: false
        schema:
          $ref: "#/definitions/matchingParams_1"
      - name: "x-access-token"
        in: "header"
        description: "user token to check if user authenticated"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Product"
        400:
          description: "Validation Error"
          schema:
            type: "string"
        403:
          description: "Forbidden"
          schema:
            type: "string"
        500:
          description: "Internal Server Error"
          schema:
            type: "string"
      security:
      - loanstore_auth:
        - "write:loans"
        - "read:loans"
      x-swagger-router-controller: "Product"
securityDefinitions:
  loanstore_auth:
    type: "oauth2"
    authorizationUrl: "https://tasaheel.swagger.io/oauth/authorize"
    flow: "implicit"
    scopes:
      write:loans: "modify loans in your account"
      read:loans: "read your loans"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
...
  Product:
    type: object
    properties:
      description:
        type: string
      categoryCode:
        type: integer
      subCategoryCode:
        type: integer
      productCode:
        type: integer
      price:
        type: integer
        format: float
      productName:
        type: string
      installmentOptions:
        type: array
        items:
          $ref: "#/definitions/installmentOptions"
      img:
        type: string
  installmentOptions:
    type: object
    properties:
      numberOfMonths:
        type: integer
      deposit:
        type: number
        format: float
      monthlyInstallment:
        type: number
        format: float
...
  matchingParams_1:
    type: "object"
    properties:
      categoryCode:
        type: "number"
      subCategoryCode:
        type: "number"